Move convertToOptionValueInTimeDropDown from TimeHelper to TimeHelperExtension
